Merge upto trunk rev 928243
The following revision of Michael Busch was left out (I do not know how this affects flex): 926791 (buschmi, LUCENE-2329: Use parallel arrays instead of PostingList objects in TermsHash*)

Also I added a noncommit to the Directory.copyTo() method, as the IndexFileNameFilter now needs a Codec which is unknown to the directory. I replaced with Arrays.asList(listAll()), without filter, for now.


git-svn-id: https://svn.apache.org/repos/asf/lucene/java/branches/flex_1458@928246 13f79535-47bb-0310-9956-ffa450edef68